add utils file and refactoring restoreWallet test
