+++ tests/e2e/restoreWallet.test.ts
@@ -1,0 +1,26 @@
+import { setup, importWallet } from './utils'
+
+const SEED = ['express', 'pretty', 'dinner', 'first', 'someone', 'reform', 'occur', 'food', 'dice', 'very', 'thumb', 'unfold']
+const btcAddress = 'n1AqFLX43FFK5dyzXkxjt6AXKLZ1TCniWw'
+
+jest.setTimeout(50 * 1000)
+
+describe('Restpore wallet', () => {
+  it('from 12 words and check recovered BTC address', async () => {
+    const { browser, page } = await setup()
+    await importWallet(page, SEED)
+
+    await page.waitForSelector('#btcAddress')
+
+    const recoveredBtcAddress = await page.$eval('#btcAddress', el => el.textContent)
+
+    await page.screenshot({
+        path: `tests//e2e/screenshots/restore_${new Date().getTime()}.jpg`,
+        type: 'jpeg'
+    });
+
+    await browser.close();
+
+    expect(btcAddress).toBe(recoveredBtcAddress)
+  })
+})
